requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-filter==24.2
django-guardian==2.4.0
djangorestframework==3.15.1
fonttools==4.52.4
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pdfkit==1.0.0
pillow==10.3.0
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
